Fixed a bug with the email regex. It can has multiple hashes. Also moved the gather_hash_function to ilapfuncs.py
